+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.0.4:
+- Fixed a bug when several matches are happening at the same time
+- Customized discord.ext.tasks as lib.tasks to have better flexibility on tasks
+
 # v1.0.3:
 - It is no longer possible to register with a character that is already registered
 - Team captains can now select a map